+++ src/RouterAndDataLoader.tsx
@@ -9,9 +9,9 @@ import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import useInterval from 'react-use/esm/useInterval';
 import RewardModal from './components/rewards/RewardModal';
 import DevelopmentOnlyMenu from './components/ui/DevelopmentOnlyMenu';
+import { GlobalSnackbar } from './components/ui/GlobalSnackbar';
 import NavBar from './components/ui/NavBar/NavBar';
 import Sidebar from './components/ui/Sidebar';
-import { GlobalSnackbar } from './components/unsorted/GlobalSnackbar';
 import { ExpirienceProgressBar } from './components/users/ExpirienceProgressBar';
 import FAQPage from './pages/FAQPage';
 import { FocusModePage } from './pages/FocusModePage';
